requirements.txt
fastapi
uvicorn
nltk
psycopg2-binary
requests
